+++ features/step_definitions/json_steps.ts
@@ -134,6 +134,20 @@ Then(
   }
 );
 
+Then("the JSON report should contain a spec", async function () {
+  const absolutejsonPath = path.join(this.tmpDir, "cucumber-report.json");
+
+  const jsonFile = await fs.readFile(absolutejsonPath);
+
+  const actualJsonOutput = JSON.parse(jsonFile.toString());
+
+  if (actualJsonOutput.length !== 1) {
+    throw new Error(
+      `Expected to find a single spec, but found ${actualJsonOutput.length}`
+    );
+  }
+});
+
 Then("the JSON report shouldn't contain any specs", async function () {
   const absolutejsonPath = path.join(this.tmpDir, "cucumber-report.json");